# frontend/web/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DAIso",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/current": {
      "put": {
        "tags": [
          "projects"
        ],
        "operationId": "SetCurrentProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "operationId": "GetUsage",
        "parameters": [
          {
            "name": "grain",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tool",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProjectItem": {
        "required": [
          "name",
          "path",
          "sessionCount",
          "lastActivity",
          "exists"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "\uD3F4\uB354 \uC774\uB984."
          },
          "path": {
            "type": "string",
            "description": "\uC804\uCCB4 \uACBD\uB85C."
          },
          "sessionCount": {
            "type": "integer",
            "description": "\uC778\uB371\uC2A4\uC5D0 \uC788\uB294 \uC138\uC158 \uC218. \uCD5C\uADFC \uD3F4\uB354\uC5D0\uB9CC \uC788\uC73C\uBA74 0.",
            "format": "int32"
          },
          "lastActivity": {
            "type": [
              "null",
              "string"
            ],
            "description": "\uB9C8\uC9C0\uB9C9 \uC138\uC158\uC774 \uBC14\uB010 \uB54C. \uC138\uC158\uC774 \uC5C6\uC73C\uBA74 null.",
            "format": "date-time"
          },
          "exists": {
            "type": "boolean",
            "description": "\uD3F4\uB354\uAC00 \uC544\uC9C1 \uC788\uB294\uAC00. \uC5C6\uC73C\uBA74 \uC138\uC158\uB9CC \uB0A8\uC740 \uAC83\uC774\uB2E4."
          }
        },
        "description": "\uD504\uB85C\uC81D\uD2B8 \uD558\uB098. \uC704 \uC904 \uD504\uB85C\uC81D\uD2B8 \uC120\uD0DD\uAE30\uC640 \u0022\uBAA8\uB4E0 \uD504\uB85C\uC81D\uD2B8\u0022 \uC694\uC57D\uC774 \uC4F4\uB2E4."
      },
      "ProjectsResponse": {
        "required": [
          "projects",
          "current"
        ],
        "type": "object",
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectItem"
            },
            "description": "\uB9C8\uC9C0\uB9C9 \uC791\uC5C5\uC774 \uCD5C\uADFC\uC778 \uAC83\uBD80\uD130."
          },
          "current": {
            "type": [
              "null",
              "string"
            ],
            "description": "\uC9C0\uAE08 \uD504\uB85C\uC81D\uD2B8\uC758 \uACBD\uB85C. null \uC774\uBA74 \u0022\uBAA8\uB4E0 \uD504\uB85C\uC81D\uD2B8\u0022."
          }
        },
        "description": "\uD504\uB85C\uC81D\uD2B8 \uBAA9\uB85D\uACFC \uC9C0\uAE08 \uACE0\uB978 \uD504\uB85C\uC81D\uD2B8."
      },
      "SetCurrentProjectRequest": {
        "required": [
          "path"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": [
              "null",
              "string"
            ],
            "description": "\uBAA9\uB85D\uC5D0 \uC788\uB294 \uACBD\uB85C. null \uC774\uBA74 \u0022\uBAA8\uB4E0 \uD504\uB85C\uC81D\uD2B8\u0022."
          }
        },
        "description": "\uC9C0\uAE08 \uD504\uB85C\uC81D\uD2B8\uB97C \uBC14\uAFB8\uB294 \uC694\uCCAD."
      },
...
      "UsageBucket": {
        "required": [
          "start",
          "tokens"
        ],
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "description": "\uCE78\uC758 \uCCAB\uB0A0.",
            "format": "date"
          },
          "tokens": {
            "description": "\uADF8 \uCE78\uC758 \uD569.",
            "$ref": "#/components/schemas/UsageTokens"
          }
        },
        "description": "\uADF8\uB798\uD504 \uD55C \uCE78. \uD558\uB8E8\u00B7\uD55C \uC8FC(\uC6D4\uC694\uC77C \uC2DC\uC791)\u00B7\uD55C \uB2EC(1\uC77C \uC2DC\uC791)."
      },
      "UsageResponse": {
        "required": [
          "firstDay",
          "today",
          "totals",
          "buckets",
          "byProject",
          "byModel"
        ],
        "type": "object",
        "properties": {
          "firstDay": {
            "type": [
              "null",
              "string"
            ],
            "description": "\uAE30\uB85D\uC774 \uC788\uB294 \uCCAB\uB0A0. \uAE30\uB85D\uC774 \uC5C6\uC73C\uBA74 null.",
            "format": "date"
          },
          "today": {
            "type": "string",
            "description": "\uC624\uB298. \uB0A0\uC9DC\uB294 UTC \uAE30\uC900\uC774\uB2E4. \uC138\uC158 \uAE30\uB85D\uC758 \uB0A0\uC9DC\uAC00 UTC \uB85C \uC801\uD78C\uB2E4.",
            "format": "date"
          },
          "totals": {
            "description": "\uC624\uB298 \u00B7 \uCD5C\uADFC 7\uC77C \u00B7 \uCD5C\uADFC 30\uC77C \uD569.",
            "$ref": "#/components/schemas/UsageTotals"
          },
          "buckets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageBucket"
            },
            "description": "\uADF8\uB798\uD504 \uCE78. \uC624\uB298\uC774 \uB4E0 \uCE78\uAE4C\uC9C0 \uBE48 \uCE78 \uC5C6\uC774 \uC774\uC5B4\uC9C4\uB2E4."
          },
          "byProject": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageShare"
            },
            "description": "\uB9CE\uC774 \uC4F4 \uD504\uB85C\uC81D\uD2B8 10\uAC1C."
          },
          "byModel": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageShare"
            },
            "description": "\uBAA8\uB378 \uC804\uBD80. \uB9CE\uC774 \uC4F4 \uAC83\uBD80\uD130."
          }
        },
        "description": "\uC0AC\uC6A9\uB7C9 \uD0ED \uD55C \uD654\uBA74."
      },
      "UsageShare": {
        "required": [
          "key",
          "label",
          "tokens",
          "share"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "\uD504\uB85C\uC81D\uD2B8 \uACBD\uB85C \uB610\uB294 \uBAA8\uB378 \uC774\uB984."
          },
          "label": {
            "type": "string",
            "description": "\uBCF4\uC5EC \uC904 \uC774\uB984. \uD504\uB85C\uC81D\uD2B8\uB294 \uD3F4\uB354 \uC774\uB984\uC774\uACE0, \uACB9\uCE58\uBA74 \uC0C1\uC704 \uD3F4\uB354\uAE4C\uC9C0 \uBD99\uC778\uB2E4."
          },
          "tokens": {
            "description": "\uD569.",
            "$ref": "#/components/schemas/UsageTokens"
          },
          "share": {
            "type": "number",
            "description": "\uC804\uCCB4\uC5D0\uC11C \uCC28\uC9C0\uD558\uB294 \uBE44\uC728. 0~1.",
            "format": "double"
          }
        },
        "description": "\uD504\uB85C\uC81D\uD2B8\uBCC4\u00B7\uBAA8\uB378\uBCC4 \uD55C \uC904."
      },
      "UsageTokens": {
        "required": [
          "input",
          "output",
          "cacheCreate",
          "cacheRead",
          "total"
        ],
        "type": "object",
        "properties": {
          "input": {
            "type": "integer",
            "format": "int64"
          },
          "output": {
            "type": "integer",
            "format": "int64"
          },
          "cacheCreate": {
            "type": "integer",
            "format": "int64"
          },
          "cacheRead": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        },
        "description": "\uD1A0\uD070 \uB124 \uAC08\uB798\uC640 \uD569."
      },
      "UsageTotals": {
        "required": [
          "today",
          "last7Days",
          "last30Days"
        ],
        "type": "object",
        "properties": {
          "today": {
            "$ref": "#/components/schemas/UsageTokens"
          },
          "last7Days": {
            "$ref": "#/components/schemas/UsageTokens"
          },
          "last30Days": {
            "$ref": "#/components/schemas/UsageTokens"
          }
        },
        "description": "\uCD5C\uADFC \uAE30\uAC04 \uD569."
      }
    }
  },
...
}